Collapse dtos in one file
