--- app/dto/create-device.dto.ts
@@ -1,6 +1,0 @@
-export class CreateDeviceDTO {
-    readonly name!: string;
-    readonly macaddress!: string;
-    readonly topic!: string;
-    readonly description!: string;  
-}
