--- app/dto/create-user.dto.ts
@@ -1,6 +1,0 @@
-export class CreateUserDTO {
-    readonly username!: string;
-    readonly fullname!: string;
-    readonly email!: string;
-    readonly phone!: string;
-}
