+++ app/dto/user.dto.ts
@@ -5,3 +5,10 @@ export class UserDTO {
     readonly email!: string;
     readonly phone!: string;
 }
+
+export class CreateUserDTO {
+    readonly username!: string;
+    readonly fullname!: string;
+    readonly email!: string;
+    readonly phone!: string;
+}
